Remove more stuff
