+++ client/network-gossip/src/bridge.rs
@@ -21,10 +21,8 @@ use crate::{
 	Network, Syncing, Validator,
 };
 
-use sc_network_common::{
-	protocol::{event::Event, ProtocolName},
-	sync::SyncEvent,
-};
+use sc_network::{event::Event, types::ProtocolName};
+use sc_network_common::sync::SyncEvent;
 use sc_peerset::ReputationChange;
 
 use futures::{
@@ -344,7 +342,7 @@ mod tests {
 		config::MultiaddrWithPeerId, NetworkBlock, NetworkEventStream, NetworkNotification,
 		NetworkPeers, NotificationSenderError, NotificationSenderT as NotificationSender,
 	};
-	use sc_network_common::{protocol::role::ObservedRole, sync::SyncEventStream};
+	use sc_network_common::{role::ObservedRole, sync::SyncEventStream};
 	use sp_runtime::{
 		testing::H256,
 		traits::{Block as BlockT, NumberFor},
