+++ client/network-gossip/src/lib.rs
@@ -68,8 +68,10 @@ pub use self::{
 };
 
 use libp2p::{multiaddr, PeerId};
-use sc_network::{NetworkBlock, NetworkEventStream, NetworkNotification, NetworkPeers};
-use sc_network_common::{protocol::ProtocolName, sync::SyncEventStream};
+use sc_network::{
+	types::ProtocolName, NetworkBlock, NetworkEventStream, NetworkNotification, NetworkPeers,
+};
+use sc_network_common::sync::SyncEventStream;
 use sp_runtime::traits::{Block as BlockT, NumberFor};
 use std::iter;
