+++ client/network/src/event.rs
@@ -19,10 +19,12 @@
 //! Network event types. These are are not the part of the protocol, but rather
 //! events that happen on the network like DHT get/put results received.
 
-use super::ProtocolName;
-use crate::protocol::role::ObservedRole;
+use crate::types::ProtocolName;
+
 use bytes::Bytes;
 use libp2p::{core::PeerId, kad::record::Key};
+
+use sc_network_common::role::ObservedRole;
 
 /// Events generated by DHT as a response to get_value and put_value requests.
 #[derive(Debug, Clone)]
